Working banana and rectangle svg.
Use banana and rectangle in svg format to change its color using js.
png format doesnt work. The home ejs file renders the slider that has 32
colors on it, along with it are the banana and rectangle svg that change
their color evertime the page is refreshed. These colors range from rgb(255, 255, 0)
to rgb(255, 159, 0). The slider has colors that are divisible by 3 on it.